Update non-const Dictionary to be `readonly`
Update the non constant Dictionary `TURN` to be readonly and achieve
what was intended by making it `const` which was not applicable in this
scenario since it had to be constructed first using `new`. This will
effectively prevent modifications to the Dictionary during regular use.,
since it does not need to change and need to stay consistent.

== References:
- [Proper way to initialize a C\# dictionary with values already in it?
  \- Stack Overflow]
  (https://stackoverflow.com/questions/17047602/proper-way-to-initialize-a-c-sharp-dictionary-with-values-already-in-it)

- [readonly \(C\# Reference\) \| Microsoft Docs]
  (https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/readonly)

- [c\# \- The expression being assigned to '\.\.\.\.' must be constant
  \- Stack Overflow]
  (https://stackoverflow.com/questions/19222533/the-expression-being-assigned-to-must-be-constant)