Add `exec` method to Robot class
Add the execute method `exec` to the Robot class in order to process
incoming string command and command arguments on the current Robot
instance. This allows outside input from a User to call methods on the
Robot as String value equivalent; e.g.

    robot.exec("MOVE") == robot.move()

This will be later used to pass User input from either Console Input or
a File, line by line into the Robot to update its current position.

== References:
- [C\# String Switch Examples \- Dot Net Perls]
  (https://www.dotnetperls.com/string-switch)

- [C\# Switch Statement]
  (https://www.tutorialspoint.com/csharp/switch_statement_in_csharp.htm)